requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.5
httpx==0.28.1
starlette==0.46.2
fastapi-cache2==0.2.2
